// src/WEBL/ClientApp/src/app/core/models/stockItem.models.ts
export class stockItem {
  id: number;
  code: string;
  productName: string;
  internalProductName: string;
  uomid: number;
  stockCategoryId: number;
  minThreshold: number;
  quantity: number;
  price: number;
  priceLookUpPrice: number;
  uomName: string;
  categoryName: string;
  stockGroupId: number;
  storageTypeId: number;
  monitored: boolean;
  storeTypeId: number;
  supplierId: number;
  currentPrice: number;
  defaultDepartmentId: number;
  defaultLocationId: number;
  defaultStoreId: number;
  packSize: number;
  packQuantity: number;
  itemQuantity: number;
  itemPrice: number;
  secondaryUomid: number;
  comparisonSecondValue: number;
  calculatedRatio: number;
  totalweight: number;
  deductedValue: number;
  currency: string;
}